+++ app/components/ui/parent-thread.tsx
@@ -1,5 +1,5 @@
 import { Link } from '@remix-run/react'
-import { type Post } from '#app/api-types.ts'
+import { type Post } from '#app/types/api-types.ts'
 import { PostContent } from './post-content.tsx'
 
 export function ParentThread({
