+++ app/post.ts
@@ -5,11 +5,11 @@ import {
 	type StatsPost,
 	type Post,
 	type PostWithOSizeAndScore,
-} from '#app/api-types.ts'
-import { type DBPost } from '#app/db/types.ts'
+} from '#app/types/api-types.ts'
+import { type DBPost } from '#app/types/db-types.ts'
 import { invariant } from '#app/utils/misc.tsx'
 import { vote } from '#app/vote.ts'
-import { type DB } from './db/kysely-types.ts'
+import { type DB } from './types/kysely-types.ts'
 import { checkIsAdminOrThrow } from './utils/auth.server.ts'
 
 export async function createPost(
