+++ app/routes/index.tsx
@@ -2,7 +2,7 @@ import { type LoaderFunctionArgs } from '@remix-run/node'
 import { Link, useLoaderData } from '@remix-run/react'
 import moment from 'moment'
 import { useState } from 'react'
-import { type FrontPagePost } from '#app/api-types.ts'
+import { type FrontPagePost } from '#app/types/api-types.ts'
 import { Markdown } from '#app/components/markdown.tsx'
 import { Button } from '#app/components/ui/button.tsx'
 import { PostContent } from '#app/components/ui/post-content.tsx'
