+++ app/routes/vote.tsx
@@ -1,5 +1,5 @@
 import { type ActionFunctionArgs } from '@remix-run/node'
-import { Direction } from '#app/api-types.ts'
+import { Direction } from '#app/types/api-types.ts'
 import { db } from '#app/db.ts'
 import { getCommentTreeState } from '#app/ranking.ts'
 import { requireUserId } from '#app/utils/auth.server.ts'
