+++ app/utils/auth.server.ts
@@ -2,7 +2,7 @@ import { createId } from '@paralleldrive/cuid2'
 import { redirect } from '@remix-run/node'
 import bcrypt from 'bcryptjs'
 import { safeRedirect } from 'remix-utils/safe-redirect'
-import { type DBPassword, type DBUser } from '#app/db/types.ts'
+import { type DBPassword, type DBUser } from '#app/types/db-types.ts'
 import { db } from '#app/db.ts'
 import { combineHeaders, invariant } from './misc.tsx'
 import { authSessionStorage } from './session.server.ts'
