+++ app/vote.ts
@@ -1,8 +1,8 @@
 import { type Transaction, sql } from 'kysely'
-import { Direction, type VoteState } from '#app/api-types.ts'
-import { type DBVoteEvent, type DBInsertableVoteEvent } from '#app/db/types.ts'
+import { Direction, type VoteState } from '#app/types/api-types.ts'
+import { type DBVoteEvent, type DBInsertableVoteEvent } from '#app/types/db-types.ts'
 import { sendVoteEvent } from '#app/globalbrain.ts'
-import { type DB } from './db/kysely-types.ts'
+import { type DB } from './types/kysely-types.ts'
 import { invariant } from './utils/misc.tsx'
 
 export function defaultVoteState(postId: number): VoteState {
